Fix: Rebased on last branch version and fixed some code duplicated.
